rename: setActiveList 함수 분리
